requirements.txt
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyQt5==5.10
sip==4.19.7
